revert the change of move main.py to __main__.py
